+++ Assets/scripts/StatusEffect/statusEffect.cs
@@ -19,7 +19,7 @@ public abstract class StatusEffect : ScriptableObject
 
     [HideInInspector] public GameObject target;
     [HideInInspector] public GameObject source;
-    [HideInInspector] public GameObject projectile;
+    [HideInInspector] public Vector2 location;
     [HideInInspector] public int currentStacks = 0;
     [HideInInspector] public float potencyMultiplier = 1f;